Return the current book for empty query searches
Legacy controllers have onPlayFromSearch method that is called for play
button and expected to play something meaningful. For media3 sessions it
delegates to getSearchResult.
In this app search was already returning the last position for calls
with null as query, but in the mentioned scenario it is called with an
empty string, so, instead of manually loading the last book played it
is now returned for empty query searches