Redesign the chapter selection (#2825)
* feat(player): use bottomsheet for chapter selection

Switches the chapter selection dialog to use a bottomsheet. This provides more breathing room for the displayed chapters, which is
especially helpful for longer titles.

Closes: https://github.com/PaulWoitaschek/Voice/issues/2265

* feat(SelectChapterDialog): show previous chapter as first visible item

Modifies to scroll behaviour, such that the first visible item is the previous chapter, indicating to the user that the sheet is scrollable
in both directions.

* feat(selectchapterdialog): mark current chapter with highlight

Marks the currently playing chapter by applying a background color, replacing the previously used icon to more clearly indicate which
chapter is selected.

* feat(SelectChapterDialog): display start time for each chapter

* style: address lints

* refactor(SelectChapterDialog): extract backgroundColor to variable

* Skip the partially expanded state.

* Change a comment to trigger a new pipeline to re-trigger the evaluation of status checks after change

---------

Co-authored-by: Paul Woitaschek <woitaschek@gmail.com>